refactor(consensus): return commits-per-schedule as u32
The value is stored as a u32 protocol param and used as a commit-count
window, so the u64 accessor only forced narrowing casts back to u32 at
the call sites. Return u32 throughout and drop the casts.

+++ crates/iota-protocol-config/src/lib.rs
@@ -1798,12 +1798,12 @@ impl ProtocolConfig {
         self.feature_flags.enable_pcool_flow
     }
 
-    pub fn commits_per_schedule(&self) -> u64 {
+    pub fn commits_per_schedule(&self) -> u32 {
         if cfg!(msim) {
             // Exercise faster leader-schedule rotation in simtests.
-            min(10, self.consensus_commits_per_schedule.unwrap_or(300)) as u64
+            min(10, self.consensus_commits_per_schedule.unwrap_or(300))
         } else {
-            self.consensus_commits_per_schedule.unwrap_or(300) as u64
+            self.consensus_commits_per_schedule.unwrap_or(300)
         }
     }
 }
